+++ src/sim/models/Twitter.scala
@@ -4,9 +4,9 @@ object Twitter extends SocialNetwork
 {
 
   override val name = "twitter"
-  var nbSub = 0
+  var nbSub = 0.
 
-  def count(nb: Int) {
+  def count(nb: Double) {
 		nbSub += nb
 	}
 }
