+++ src/sim/models/Youtube.scala
@@ -3,9 +3,9 @@ package sim.models
 object Youtube extends SocialNetwork
 {
   override val name = "youtube"
-  var nbSub = 0
+  var nbSub = 0.
 
-  def count(nb: Int) {
+  def count(nb: Double) {
 		nbSub += nb
 	}
 }
